pyproject.toml
[tool.poetry]
name = "resume-scorer"
version = "0.1.0"
description = ""
authors = ["Kanha Upadhyay <kanha.upadhyay@sifars.com>"]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.31.1"
openai = "^1.12.0"
pdf2image = "^1.17.0"
pytesseract = "^0.3.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
